requirements.txt
django
django-cors-headers
djangorestframework
drf-spectacular
transformers
torch
datasets
python-dotenv
cryptography
redis 
celery
flower
loguru